optmizing client component and add restriction nav

+++ src/app/components/cliente/cliente-delete/cliente-delete.component.ts
@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
 import { ClienteService } from "src/app/services/cliente.service";
 import { Cliente } from "src/app/models/cliente";
 import { ToastrService } from "ngx-toastr";
@@ -7,9 +7,9 @@ import { ActivatedRoute, Router } from "@angular/router";
 @Component({
   selector: "app-cliente-delete",
   templateUrl: "./cliente-delete.component.html",
-  styleUrl: "./cliente-delete.component.css",
+  styleUrls: ["./cliente-delete.component.css"],
 })
-export class ClienteDeleteComponent {
+export class ClienteDeleteComponent implements OnInit {
   cliente: Cliente = {
     id: "",
     nome: "",
@@ -28,36 +28,40 @@ export class ClienteDeleteComponent {
   ) {}
 
   ngOnInit(): void {
-    this.cliente.id = this.route.snapshot.paramMap.get("id");
+    this.cliente.id = this.route.snapshot.paramMap.get("id")!;
     this.findById();
   }
 
   findById(): void {
-    this.service.findById(this.cliente.id).subscribe((response) => {
-      this.cliente = response;
-    });
+    this.service.findById(this.cliente.id).subscribe(
+      (response) => {
+        this.cliente = response;
+      },
+      (ex) => {
+        this.toast.error("Erro ao carregar cliente: " + ex.error.message);
+      }
+    );
   }
 
   delete(): void {
-    // Convertendo os perfis para números
-    this.cliente.perfis = this.cliente.perfis.map((perfil) => Number(perfil));
-
-    this.service.delete(this.cliente.id).subscribe(
-      () => {
-        this.toast.success("Cliente deletado com sucesso", "Deletado");
-        this.router.navigate(["clientes"]);
-      },
-      (ex) => {
-        if (ex.error.errors) {
-          ex.error.errors.forEach(
-            (element: { message: string | undefined }) => {
-              this.toast.error(element.message || "Erro desconhecido");
-            }
-          );
-        } else {
-          this.toast.error(ex.error.message);
+    if (confirm("Tem certeza que deseja deletar este cliente?")) {
+      this.service.delete(this.cliente.id).subscribe(
+        () => {
+          this.toast.success("Cliente deletado com sucesso", "Deletado");
+          this.router.navigate(["clientes"]);
+        },
+        (ex) => {
+          if (ex.error.errors) {
+            ex.error.errors.forEach(
+              (element: { message: string | undefined }) => {
+                this.toast.error(element.message || "Erro desconhecido");
+              }
+            );
+          } else {
+            this.toast.error("Erro ao deletar cliente: " + ex.error.message);
+          }
         }
-      }
-    );
+      );
+    }
   }
 }
